feat(InlineCode): Add component
This also corrects a line height issue that caused lines with inline
code to look droopy next to the numbering of an ordered list.
