+++ .loki/reference/chrome_laptop_MdxProvider_Lists.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d855daa1f4a2684824f82d65a64d66bc27ee47dd0b090e50d702b30af2ce74b5
-size 66790
+oid sha256:db10209b449702b47a18a4a21ba8074de13bdc8974186a7b97a09cecaaadf27e
+size 71263
