+++ src/index.ts
@@ -1,4 +1,5 @@
 export { CodeBlock } from './components/CodeBlock';
+export { InlineCode } from './components/InlineCode';
 export { InternalLink } from './components/InternalLink';
 export { MdxProvider } from './components/MdxProvider';
 export { SmartTextLink } from './components/SmartTextLink';
