+++ src/stories/markdowns/lists.mdx
@@ -1,3 +1,8 @@
+1. 1\. Plain
+2. 2\. With `inline code`
+
+---
+
 1. First item
 2. Second item
 3. Third item
